fix: Move TimeZone, TimeDelta types to common/configs
